backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.23
pydantic==2.4.2
python-dotenv==1.0.0
google-generativeai==0.3.2
httpx==0.25.1